remove battleUIFragment
